Fix it doesn't compile for x64.
git-svn-id: https://subversion.assembla.com/svn/q3/trunk@2446 8af8166b-12a6-a448-a533-9086ace3f9f6

+++ q3/qs/src/util/process.cpp
@@ -170,7 +170,7 @@ int qs::Process::exec(const WCHAR* pwszCommand,
 		listHandle.push_back(hThreadStdout);
 	if (hThreadStderr)
 		listHandle.push_back(hThreadStderr);
-	::WaitForMultipleObjects(listHandle.size(), &listHandle[0], TRUE, INFINITE);
+	::WaitForMultipleObjects(static_cast<DWORD>(listHandle.size()), &listHandle[0], TRUE, INFINITE);
 	
 	for (HandleList::const_iterator it = listHandle.begin() + 1; it != listHandle.end(); ++it) {
 		DWORD dwExitCode = 0;
